refactor(app): move the workspace store, indicators and layer wiring out
`app.rs` was 7,499 production lines after PR #295 and still the largest
ceiling in the size baseline -- the first file every agent reads to change
anything in the window. Three cohesive `impl QuantickApp` groups leave it,
the same way the demo hooks, drawing input, health and workspace restore
left in #295:

  app/workspace_save.rs      (1,136) the write half of the workspace round
                                     trip whose read half is workspace_restore
  app/indicator_manager.rs     (835) legends, settings, presets, script files,
                                     attach and detach
  app/chart_layers_wiring.rs   (270) the layer actions, the mask and the
                                     restore

`app.rs` falls to 5,339 production lines, its ceiling is tightened to match
and the `!budget` falls 2,160. None of the three new files reaches the
1,500-line threshold, so none takes a baseline entry and the fall is a trade
of nothing for 2,160.

The bodies are unchanged. Of 2,274 added lines, 1,812 are verbatim moves and
46 differ only by a new `pub(super)` -- a method a child module owns and
`app.rs` still calls has to be visible to it, the visibility the five
existing siblings already use. The remaining 87 are module doc comments,
import blocks, `impl QuantickApp {` wrappers, three `mod` lines and the
baseline's own prose; `docs/evidence/.../move-audit.txt` lists every one.

`SCRIPT_RELOAD_POLL_INTERVAL` travels with `poll_script_files`, its only
reader. No hook read or declaration moves: the `declare_hooks!` slice stays
whole in `app.rs`, and the generated registry and capability inventory are
byte-identical.

Five lines are added to `app/tests/mod.rs`. `CandlePreset` and
`IndicatorEvent` reached the twelve test files through `app.rs`'s imports,
and the indicator manager took the last production reader of each away;
with `warnings = "deny"` the import cannot stay and the tests still need the
name. `tests/mod.rs` already binds `DrawingsDemo` for this exact reason
after #295, and these sit beside it. No test body, name or assertion
changes, and the suite runs the same 1,894 tests.
