+++ crates/app/src/app.rs
@@ -18,33 +18,41 @@ use eframe::egui;
 
 use crate::canvas_layout::{MAX_CANVAS_PANES, PaneIdAllocator};
 
+mod chart_layers_wiring;
 mod demo_hooks;
 mod drawing_input;
 mod health;
+mod indicator_manager;
 mod layout_wiring;
 mod workspace_restore;
-use crate::chart_layers::{self, ChartLayer};
+mod workspace_save;
+use crate::chart_layers;
+// `ChartLayer` left this file's production code with `maintain_chart_layers`
+// and lives in `app::chart_layers_wiring` now. The one name still read here is
+// read by `heatmap_lamp_on`, which is itself test-only, so the import is gated
+// the same way rather than kept alive by a `chart_layers::` prefix that would
+// appear on exactly one line of the file.
+#[cfg(test)]
+use crate::chart_layers::ChartLayer;
 use crate::config::AppConfig;
 use crate::dock::{Dock, DockEnv, DockTab};
 use crate::drawings::{self, DeleteOutcome, DrawingAuthor};
 use crate::feed_notice;
 use crate::harness::{ContextMenuPane, Harness, ScriptedMenu};
-use crate::indicator_legend;
-use crate::indicator_panel::{self, SettingsDialog, SettingsOutcome};
-use crate::indicator_worker::{IndicatorCommand, IndicatorEvent, IndicatorSource, SlotId};
-use crate::indicators::IndicatorView;
+use crate::indicator_panel::SettingsDialog;
+use crate::indicator_worker::{IndicatorSource, SlotId};
 use crate::indicators::library::ScriptLibrary;
 use crate::indicators::preset_file;
-use crate::indicators::state_file::{self, SavedInput, SavedKind};
+use crate::indicators::state_file::{self, SavedKind};
 use crate::loading::{self, LoadingScope, LoadingTask};
 use crate::metrics::{self, FrameStats};
 use crate::pane::{self, ChartPane, DRAWING_ANCHOR_RADIUS_PX, PaneSide};
 use crate::replay_view::{ReplayAction, ReplayView};
 use crate::state::BarSpec;
 use crate::statusbar;
-use crate::style::{CandlePreset, ChartStyle};
+use crate::style::ChartStyle;
 use crate::symbols_file::{self, AddedSymbols};
-use crate::tab::{CanvasChrome, CanvasLayout, LegendFold, Tab};
+use crate::tab::{CanvasChrome, CanvasLayout, Tab};
 use crate::tabstrip::{self, TabAction};
 use crate::theme;
 use crate::timezone::TzOffset;
@@ -52,7 +60,7 @@ use crate::toolbar::{self, ToolbarAction};
 use crate::toolrail::{Tool, ToolRail, ToolboxDock};
 use crate::ui_state;
 use crate::window_scale;
-use crate::workspace_store::{LayoutStore, StorePaths, WorkspacePick, WorkspaceStore};
+use crate::workspace_store::{LayoutStore, StorePaths, WorkspaceStore};
 use quantick_feed::history_reach;
 use quantick_feed::{self as feed, FeedCommand, FeedHandle, ReplayControl};
 use quantick_orderflow::LaneWindow;
@@ -71,8 +79,6 @@ const DEMO_VISIBLE_SLOTS: usize = 90;
 /// deterministic pair — a native added later must not silently change what
 /// every existing capture shows.
 const AUTOSTART_NATIVES: &[&str] = &["native.ema", "native.cvd"];
-/// How often the hot-reload poll checks script files for changes.
-const SCRIPT_RELOAD_POLL_INTERVAL: Duration = Duration::from_millis(1_000);
 /// Horizontal offset of a duplicated drawing, so the copy is visibly a copy.
 const DUPLICATE_OFFSET_BARS: f32 = 2.0;
 
@@ -1668,73 +1674,6 @@ impl QuantickApp {
         self.active_tab
     }
 
-    /// Put one Quantick Pine script on the focused pane behind a fresh slot.
-    ///
-    /// The one door: the script library's click arrives here, and so does an
-    /// authorized agent's `indicator.script.attach`. Returns the tab, the
-    /// pane and the slot, which is what a caller needs to detach it again.
-    pub(crate) fn attach_script_indicator(
-        &mut self,
-        name: String,
-        text: String,
-        by_operator: bool,
-    ) -> (u64, crate::control::PaneSideDto, SlotId) {
-        let slot = self
-            .focused_pane_mut()
-            .add_indicator(IndicatorSource::Script {
-                name: name.clone(),
-                text,
-            });
-        let owner = self.target_slot(slot);
-        let kind = SavedKind::Script { name };
-        self.slot_kinds.push((owner, kind.clone()));
-        // Whose slot this is decides who may take it away again: the annotate
-        // tier removes what it attached, never what the trader put there. An
-        // operator's overlay stays on the one pane it was attached to and
-        // out of the layout; the trader's script is a layout edit and goes
-        // onto every pane.
-        if by_operator {
-            self.operator_slots.insert(owner);
-        } else {
-            self.mirror_add(owner, &kind);
-        }
-        self.note_indicator_edit_at(owner.tab, owner.side);
-        (owner.tab, owner.side.into(), slot)
-    }
-
-    /// Take one slot **an operator attached** off the chart, through the same
-    /// removal the indicator legend's own button uses.
-    ///
-    /// `Err` when the slot is one the trader put there: this tier adds and
-    /// takes back its own, and never removes work done by hand (plan §2.6).
-    /// `Ok(false)` when there is no such slot at all.
-    pub(crate) fn detach_script_indicator(&mut self, slot: u64) -> Result<bool, ()> {
-        // A slot number is allocated per pane, so several panes can carry the
-        // same one: the operator's own is the one to take, and matching on the
-        // number alone would remove whichever pane happened to be registered
-        // first — the trader's, as often as not.
-        let mut known = false;
-        let mut target = None;
-        for (owner, _) in &self.slot_kinds {
-            if owner.slot.0 != slot {
-                continue;
-            }
-            known = true;
-            if self.operator_slots.contains(owner) {
-                target = Some(*owner);
-                break;
-            }
-        }
-        let Some(target) = target else {
-            // A slot that exists but belongs to the trader is refused; one
-            // that exists nowhere simply was not there.
-            return if known { Err(()) } else { Ok(false) };
-        };
-        self.remove_indicator_at(target);
-        self.operator_slots.remove(&target);
-        Ok(true)
-    }
-
     /// Open the assistant's popup. One at a time: a second message replaces
     /// the first rather than stacking windows over a chart someone is
     /// trading, and the trader dismisses it.
@@ -2653,2103 +2592,6 @@ impl QuantickApp {
                 .market(quantick_engine::Side::Sell),
             ToolbarAction::PaperClose => self.active_tab_mut().paper.close_position(),
         }
-    }
-
-    /// Open the dialog for whichever indicator a gesture on a pane asked
-    /// about: the pane header, a collapsed strip, or an overlay's own line.
-    ///
-    /// Drained here rather than acted on where it was read, because the dialog
-    /// belongs to the window and the gestures are read deep inside a pane's
-    /// input pass. Both sides of a split are asked, and each request resolves
-    /// against the pane that raised it — the legend's rule (MAJOR-4), applied
-    /// to the same problem one layer down.
-    fn open_requested_indicator_settings(&mut self) {
-        let tab_id = self.active_tab().id;
-        // The harness hook waits for the view it names, exactly as the restored
-        // hidden flags do: the indicator is born from the worker's first
-        // Rebuilt, which is several frames after the window opens.
-        if let Some((index, tab)) = self.harness.settings_autostart() {
-            // The focused pane first, then the flow pane. A split tab can open
-            // with the *time* pane focused while every indicator — the ones the
-            // other autostart hook adds, and the ones the state file restores —
-            // lives on the flow pane, and a hook that only asked the focused
-            // side then waited for a view that was never coming. Silence is the
-            // worst failure a validation hook can have: the run captures a
-            // chart with no dialog on it and nothing says why.
-            let focused = self.active_tab().focused_side();
-            let found = [focused, PaneSide::Flow].into_iter().find_map(|side| {
-                self.tabs
-                    .iter()
-                    .find(|candidate| candidate.id == tab_id)
-                    .map(|candidate| candidate.pane(side))
-                    .and_then(|pane| pane.indicators.all().get(index))
-                    .map(|view| (side, view.slot))
-            });
-            if let Some((side, slot)) = found {
-                self.harness.settings_autostart_opened();
-                self.open_indicator_settings_at(TabSlot {
-                    tab: tab_id,
-                    side,
-                    slot,
-                });
-                if let Some(dialog) = self.indicator_settings.as_mut() {
-                    dialog.tab = tab;
-                }
-            }
-        }
-        let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == tab_id) else {
-            return;
-        };
-        let requests: SmallVec<[(PaneSide, SlotId); MAX_CANVAS_PANES]> = tab
-            .panes_with_sides_mut()
-            .filter_map(|(pane, side)| pane.take_settings_request().map(|slot| (side, slot)))
-            .collect();
-        for (side, slot) in requests {
-            {
-                self.open_indicator_settings_at(TabSlot {
-                    tab: tab_id,
-                    side,
-                    slot,
-                });
-            }
-        }
-    }
-
-    /// Fold or unfold one pane's indicator legend.
-    ///
-    /// The one place the state changes. The chevron, the menu entry, the
-    /// hotkey and the harness hook all arrive here with the outcome they want,
-    /// so none of them can leave a pane in a state the others would not have
-    /// produced — and an operator that is not holding the mouse names this
-    /// call rather than a click.
-    fn set_legend_collapsed(&mut self, tab_id: u64, side: PaneSide, collapsed: bool) {
-        let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == tab_id) else {
-            return;
-        };
-        tab.pane_mut(side).legend_collapsed = collapsed;
-    }
-
-    /// Whether the focused pane's legend is folded — what the menu label and
-    /// the hotkey read before deciding which way to flip it.
-    fn focused_legend_collapsed(&self) -> bool {
-        let tab = self.active_tab();
-        tab.pane(tab.focused_side()).legend_collapsed
-    }
-
-    /// Fold or unfold the legend of the pane the chrome speaks for. The menu
-    /// entry and the hotkey both act through this, so "which chart did that
-    /// affect" has one answer: the focused one, the same pane every other
-    /// chrome control acts on.
-    fn set_focused_legend_collapsed(&mut self, collapsed: bool) {
-        let tab_id = self.active_tab().id;
-        let side = self.active_tab().focused_side();
-        self.set_legend_collapsed(tab_id, side, collapsed);
-    }
-
-    /// Flip a slot's render-side eye, wherever the slot lives. Addressed by
-    /// [`TabSlot`], never by focus: the legend acts on the pane it is drawn
-    /// on, and the toolbar path builds its target from focus before calling.
-    fn toggle_indicator_hidden_at(&mut self, target: TabSlot) {
-        let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == target.tab) else {
-            return;
-        };
-        tab.pane_mut(target.side)
-            .indicators
-            .toggle_hidden(target.slot);
-        self.mirror_hidden(target);
-        self.note_indicator_edit_at(target.tab, target.side);
-    }
-
-    /// Remove a slot, wherever it lives. UI first (the entry vanishes this
-    /// frame), worker second; events already in flight for the slot are
-    /// dropped on apply.
-    fn remove_indicator_at(&mut self, target: TabSlot) {
-        if !self.tabs.iter().any(|tab| tab.id == target.tab) {
-            return;
-        }
-        // The mirrors first, while the slot's layout position can still be
-        // read off the bookkeeping.
-        self.mirror_remove(target);
-        let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == target.tab) else {
-            return;
-        };
-        let pane = tab.pane_mut(target.side);
-        pane.indicators.remove(target.slot);
-        pane.indicator_worker
-            .send(IndicatorCommand::Remove(target.slot));
-        self.slot_kinds.retain(|(owner, _)| *owner != target);
-        self.operator_slots.remove(&target);
-        self.script_files.retain(|(owner, ..)| *owner != target);
-        self.pending_hidden.retain(|owner| *owner != target);
-        self.pending_styles.retain(|(owner, _)| *owner != target);
-        self.note_indicator_edit_at(target.tab, target.side);
-    }
-
-    /// Open the settings dialog for a slot, wherever it lives.
-    fn open_indicator_settings_at(&mut self, target: TabSlot) {
-        let Some(view) = self
-            .tabs
-            .iter()
-            .find(|tab| tab.id == target.tab)
-            .map(|tab| tab.pane(target.side))
-            .and_then(|pane| {
-                pane.indicators
-                    .all()
-                    .iter()
-                    .find(|view| view.slot == target.slot)
-            })
-        else {
-            return;
-        };
-        self.indicator_settings = Some(SettingsDialog {
-            slot: target.slot,
-            title: view.label().to_owned(),
-            tab: indicator_panel::SettingsTab::default(),
-            draft: view.input_values.clone(),
-            committed: view.input_values.clone(),
-            previewed: false,
-            settled: false,
-            preset_label: None,
-            preset_name_draft: String::new(),
-        });
-        self.indicator_settings_target = target;
-    }
-
-    /// Draw each visible pane's indicator legend and run what its rows asked
-    /// for. Actions resolve against the pane the legend was drawn on — the
-    /// legend must never act on the chart beside it (the audit's MAJOR-4
-    /// trap, avoided by construction).
-    fn draw_indicator_legends(&mut self, ctx: &egui::Context) {
-        let tab_id = self.active_tab().id;
-        // Whether a context chart is on screen — from what the layout holds
-        // and whether the column is collapsed, never from one variant. Matched
-        // against `TimeAndFlow` alone, this drew no legend at all on the
-        // stacked charts of `time+time+flow`, and drew one over the flow chart
-        // against a stale rect while the column was put away.
-        let split = self.active_tab().shows_context_charts();
-        let mut pending: Vec<(PaneSide, indicator_legend::LegendAction)> = Vec::new();
-        let shown = self.active_tab().context_panes_shown();
-        let sides: SmallVec<[PaneSide; MAX_CANVAS_PANES]> = self.active_tab().sides().collect();
-        for side in sides {
-            // Only a context chart on screen draws a legend: the stack may
-            // hold a pane the layout no longer shows.
-            if let PaneSide::Time(slot) = side
-                && (!split || slot >= shown)
-            {
-                continue;
-            }
-            let pane = self.active_tab().pane(side);
-            // The rect is last frame's, like every anchor the input path
-            // reads; a pane not yet drawn has none and draws no legend.
-            let Some(mut rect) = pane.last_chart_area else {
-                continue;
-            };
-            // The position HUD owns the very corner of the pane it paints on;
-            // this legend rides just below it there, and nowhere else. That
-            // pane is the one holding the HUD anchor — the focused one, not
-            // the flow one, so a split with the time pane focused drops these
-            // chips on the *time* pane. The order-flow key stacks under this
-            // legend and reads the same offset from the same place, so the two
-            // cannot drift apart.
-            rect.min.y += indicator_legend::hud_offset_px(
-                pane.paper_hud_anchor().is_some()
-                    && self.active_tab().paper.position_summary().is_some(),
-            );
-            // The slot being live-previewed by the settings dialog, if it
-            // lives on this pane: its legend row wears a "preview" chip.
-            let preview_slot = self
-                .indicator_settings
-                .as_ref()
-                .filter(|dialog| dialog.previewed)
-                .filter(|_| {
-                    self.indicator_settings_target.tab == tab_id
-                        && self.indicator_settings_target.side == side
-                })
-                .map(|dialog| dialog.slot);
-            for action in indicator_legend::draw(
-                ctx,
-                pane.id,
-                rect,
-                pane.indicators.all(),
-                preview_slot,
-                pane.legend_collapsed,
-            ) {
-                pending.push((side, action));
-            }
-        }
-        for (side, action) in pending {
-            let at = |slot| TabSlot {
-                tab: tab_id,
-                side,
-                slot,
-            };
-            match action {
-                indicator_legend::LegendAction::ToggleHidden(slot) => {
-                    self.toggle_indicator_hidden_at(at(slot));
-                }
-                indicator_legend::LegendAction::OpenSettings(slot) => {
-                    self.open_indicator_settings_at(at(slot));
-                }
-                indicator_legend::LegendAction::Remove(slot) => {
-                    self.remove_indicator_at(at(slot));
-                }
-                indicator_legend::LegendAction::SetCollapsed(collapsed) => {
-                    self.set_legend_collapsed(tab_id, side, collapsed);
-                }
-            }
-        }
-    }
-
-    /// Draw the settings dialog and execute its outcome. Apply goes through
-    /// the worker (construct anew, replace, replay) — the same path every
-    /// input change takes, UI or not.
-    fn draw_indicator_settings(&mut self, ctx: &egui::Context) {
-        // The boot hook's deferred half: fire once a slot can actually show
-        // a dialog (see `harness::Harness::wants_indicator_settings_dialog`).
-        if self.harness.wants_indicator_settings_dialog() && self.indicator_settings.is_none() {
-            let tab_id = self.active_tab().id;
-            if let Some(slot) = self
-                .active_tab()
-                .flow_pane
-                .indicators
-                .all()
-                .iter()
-                .find(|view| !view.input_values.is_empty())
-                .map(|view| view.slot)
-            {
-                self.open_indicator_settings_at(TabSlot {
-                    tab: tab_id,
-                    side: PaneSide::Flow,
-                    slot,
-                });
-                self.harness.indicator_settings_dialog_opened();
-            }
-        }
-        let target = self.indicator_settings_target;
-        // The preset shelf for this slot's kind. A slot with no registered
-        // kind (the natives autostart hook deliberately registers none) has
-        // nowhere to save to, and the dialog hides the picker.
-        let preset_names: Option<Vec<String>> = self
-            .slot_kinds
-            .iter()
-            .find(|(owner, _)| *owner == target)
-            .map(|(_, kind)| {
-                self.indicator_presets
-                    .names_for(kind)
-                    .map(str::to_owned)
-                    .collect()
-            });
-        let outcome = {
-            let Self {
-                indicator_settings,
-                tabs,
-                ..
-            } = self;
-            let Some(dialog) = indicator_settings.as_mut() else {
-                return;
-            };
-            // The tab the dialog was opened on may have been closed under it.
-            let Some(view) = tabs
-                .iter_mut()
-                .find(|tab| tab.id == target.tab)
-                .map(|tab| tab.pane_mut(target.side))
-                .and_then(|pane| pane.indicators.view_mut(dialog.slot))
-            else {
-                // The indicator was removed under the dialog.
-                *indicator_settings = None;
-                return;
-            };
-            // Follow the live label: an applied edit can retitle the
-            // indicator (`EMA(9)` → `EMA(21)`), and a dialog that stays open
-            // must not keep announcing the old one.
-            dialog.title = view.label().to_owned();
-            // The descriptor is read while the style layer beside it is
-            // written, so the two halves are split off the same view here
-            // rather than borrowed twice.
-            let IndicatorView {
-                descriptor, style, ..
-            } = view;
-            indicator_panel::draw(
-                ctx,
-                dialog,
-                &descriptor.inputs,
-                &descriptor.plots,
-                style,
-                preset_names.as_deref(),
-            )
-        };
-        match outcome {
-            SettingsOutcome::Open => {}
-            SettingsOutcome::Close => {
-                self.revert_indicator_settings_preview();
-                self.indicator_settings = None;
-            }
-            SettingsOutcome::Apply => self.apply_indicator_settings_draft(),
-            SettingsOutcome::Preview => self.preview_indicator_settings_draft(),
-            SettingsOutcome::LoadPreset(name) => self.load_indicator_preset(name),
-            SettingsOutcome::SavePreset(name) => self.save_indicator_preset(&name),
-            SettingsOutcome::DeletePreset(name) => self.delete_indicator_preset(&name),
-            // Style is render-side: the chart already shows it and there is
-            // nothing to preview or commit, so all that is owed is the file.
-            // Deliberately unlike an input edit — it takes no rebuild and no
-            // replay, and it follows the legend's eye, which is also instant
-            // and also persisted without an Apply.
-            SettingsOutcome::StyleChanged => {
-                let target = self.indicator_settings_target;
-                self.mirror_style(target);
-                self.note_indicator_edit_at(target.tab, target.side);
-            }
-        }
-    }
-
-    /// Replace the dialog's draft with a preset (`None` = the declared
-    /// defaults) and preview it — the same nudge-and-look contract as a
-    /// slider: Apply commits, Discard reverts. A saved value whose type no
-    /// longer matches its input (the script evolved under the preset) falls
-    /// back to that input's default rather than being silently coerced.
-    fn load_indicator_preset(&mut self, name: Option<String>) {
-        let target = self.indicator_settings_target;
-        let Some(specs) = self
-            .tabs
-            .iter()
-            .find(|tab| tab.id == target.tab)
-            .map(|tab| tab.pane(target.side))
-            .and_then(|pane| {
-                pane.indicators
-                    .all()
-                    .iter()
-                    .find(|view| view.slot == target.slot)
-            })
-            .map(|view| view.descriptor.inputs.clone())
-        else {
-            return;
-        };
-        let saved: Option<Vec<Option<quantick_indicators::InputValue>>> = match &name {
-            None => Some(Vec::new()),
-            Some(name) => {
-                let Some(kind) = self
-                    .slot_kinds
-                    .iter()
-                    .find(|(owner, _)| *owner == target)
-                    .map(|(_, kind)| kind)
-                else {
-                    return;
-                };
-                self.indicator_presets
-                    .get(kind, name)
-                    // `map`, not `filter_map`: a stored cell that no longer
-                    // reads (a source whose name the dialect dropped, say)
-                    // still has to hold its index. Dropping it from the list
-                    // would slide every value after it one input to the left,
-                    // and same-typed neighbours would take each other's
-                    // settings without a word.
-                    .map(|inputs| inputs.iter().map(SavedInput::to_value).collect())
-            }
-        };
-        let Some(saved) = saved else {
-            return;
-        };
-        // The same binder the worker binds a saved state file with: one rule
-        // for what a saved value means, in one place.
-        let bound = quantick_indicators::bind_by_position(&specs, &saved);
-        if bound.kept < saved.len() {
-            // The preset on screen is not the preset that was saved. Silence
-            // here would show a chart that does not match the name above it.
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "INDICATOR_PRESET_REBOUND",
-                preset = %name.as_deref().unwrap_or(indicator_panel::DEFAULT_PRESET),
-                saved = saved.len(),
-                declared = specs.len(),
-                kept = bound.kept,
-                action = "bound_by_position",
-                "some of this preset's values no longer bind; those inputs took their defaults"
-            );
-        }
-        let draft = bound.values;
-        let label = name.unwrap_or_else(|| indicator_panel::DEFAULT_PRESET.to_owned());
-        if let Some(dialog) = self.indicator_settings.as_mut() {
-            dialog.draft = draft;
-            dialog.preset_label = Some(label);
-        }
-        self.preview_indicator_settings_draft();
-    }
-
-    /// Save the dialog's current draft under `name` for this slot's kind.
-    /// Saving is a file write, never a chart change — the draft on screen
-    /// stays exactly as it was.
-    fn save_indicator_preset(&mut self, name: &str) {
-        let target = self.indicator_settings_target;
-        let Some(kind) = self
-            .slot_kinds
-            .iter()
-            .find(|(owner, _)| *owner == target)
-            .map(|(_, kind)| kind.clone())
-        else {
-            return;
-        };
-        let Some(dialog) = self.indicator_settings.as_mut() else {
-            return;
-        };
-        let inputs: Vec<SavedInput> = dialog.draft.iter().map(SavedInput::from_value).collect();
-        if self.indicator_presets.insert(&kind, name, inputs) {
-            self.indicator_presets
-                .save(self.workspace.indicator_presets_path());
-            dialog.preset_label = Some(name.trim().to_owned());
-            dialog.preset_name_draft.clear();
-        }
-    }
-
-    /// Forget a preset of this slot's kind. The values on screen stay —
-    /// deleting a name never touches the chart.
-    fn delete_indicator_preset(&mut self, name: &str) {
-        let target = self.indicator_settings_target;
-        let Some(kind) = self
-            .slot_kinds
-            .iter()
-            .find(|(owner, _)| *owner == target)
-            .map(|(_, kind)| kind.clone())
-        else {
-            return;
-        };
-        if self.indicator_presets.remove(&kind, name) {
-            self.indicator_presets
-                .save(self.workspace.indicator_presets_path());
-            if let Some(dialog) = self.indicator_settings.as_mut()
-                && dialog.preset_label.as_deref() == Some(name)
-            {
-                dialog.preset_label = None;
-            }
-        }
-    }
-
-    /// Send the open dialog's draft to the worker and keep the dialog open
-    /// (audit M2): tuning is a nudge-and-look loop, and a dialog that dies
-    /// on every Apply makes each attempt four clicks. The slot is the one
-    /// the dialog was opened on, not whatever has focus now — clicking
-    /// Apply must not retarget the edit.
-    fn apply_indicator_settings_draft(&mut self) {
-        let target = self.indicator_settings_target;
-        let Some(dialog) = self.indicator_settings.as_mut() else {
-            return;
-        };
-        dialog.committed = dialog.draft.clone();
-        dialog.previewed = false;
-        let (slot, values) = (dialog.slot, dialog.draft.clone());
-        if let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == target.tab) {
-            tab.pane_mut(target.side)
-                .indicator_worker
-                .send(IndicatorCommand::SetInputs {
-                    slot,
-                    values: values.clone(),
-                });
-        }
-        self.mirror_inputs(target, &values);
-        self.note_indicator_edit_at(target.tab, target.side);
-    }
-
-    /// Show the draft on the chart without committing it: same worker path
-    /// as Apply, but the state file is never marked dirty — what survives a
-    /// restart is always the last Apply, never a slider mid-drag. The worker
-    /// coalesces a burst of these to its own cadence.
-    fn preview_indicator_settings_draft(&mut self) {
-        let target = self.indicator_settings_target;
-        let Some(dialog) = self.indicator_settings.as_mut() else {
-            return;
-        };
-        dialog.previewed = true;
-        let (slot, values) = (dialog.slot, dialog.draft.clone());
-        if let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == target.tab) {
-            tab.pane_mut(target.side)
-                .indicator_worker
-                .send(IndicatorCommand::SetInputs { slot, values });
-        }
-    }
-
-    /// Put the last committed values back on the chart. Close's half of the
-    /// preview contract: a dialog dismissed mid-tuning leaves no trace.
-    fn revert_indicator_settings_preview(&mut self) {
-        let target = self.indicator_settings_target;
-        let Some(dialog) = self.indicator_settings.as_ref() else {
-            return;
-        };
-        if !dialog.previewed {
-            return;
-        }
-        let (slot, values) = (dialog.slot, dialog.committed.clone());
-        if let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == target.tab) {
-            tab.pane_mut(target.side)
-                .indicator_worker
-                .send(IndicatorCommand::SetInputs { slot, values });
-        }
-    }
-
-    /// Load a library script behind a fresh slot. A file that no longer
-    /// reads or a script that no longer compiles becomes the slot's error —
-    /// shown with lines and codes, never silently dropped.
-    ///
-    /// Returns the slot it claimed, so a caller that needs to address the new
-    /// indicator (restoring saved inputs, say) does not have to guess which
-    /// one it is.
-    fn add_script_indicator(&mut self, index: usize) -> Option<SlotId> {
-        let entry = self.script_library.entries().get(index)?;
-        let name = entry.name.clone();
-        match self.script_library.read(index) {
-            Some(Ok(text)) => {
-                let (_, _, slot) = self.attach_script_indicator(name, text, false);
-                let owner = self.target_slot(slot);
-                // Watch the file so a save reloads it. Registered here, with
-                // the add, so the two cannot drift apart.
-                if let Some((_, mtime)) = self.script_library.file_info(index) {
-                    self.script_files.push((owner, index, mtime));
-                }
-                Some(slot)
-            }
-            Some(Err(message)) => {
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "INDICATOR_SCRIPT_UNREADABLE",
-                    script = %name,
-                    error = %message,
-                    action = "error_slot_shown",
-                    "cannot read an indicator script"
-                );
-                // A click that produces nothing at all is the failure this
-                // function's own doc comment rules out. The compile half of
-                // that promise runs worker-side; the read half never leaves
-                // the UI thread, so the error slot is built here, from the
-                // same two events the worker would have sent.
-                let pane = self.focused_pane_mut();
-                // The same kind string the healthy path derives from its
-                // source, so an error slot the trader fixes and reloads keeps
-                // whatever they had drawn on its pane.
-                let slot = pane.indicators.allocate_slot(format!("script.{name}"));
-                pane.indicators.apply(IndicatorEvent::Rebuilt {
-                    slot,
-                    descriptor: quantick_indicators::IndicatorDescriptor {
-                        title: name,
-                        short_title: None,
-                        overlay: false,
-                        plots: Vec::new(),
-                        fills: Vec::new(),
-                        inputs: Vec::new(),
-                    },
-                    columns: Vec::new(),
-                    bar_paint: Vec::new(),
-                    rows: 0,
-                    inputs: Vec::new(),
-                    stale: None,
-                });
-                pane.indicators.apply(IndicatorEvent::Error {
-                    slot,
-                    error: quantick_indicators::EvalError {
-                        bar_index: 0,
-                        message,
-                    },
-                });
-                Some(slot)
-            }
-            None => None,
-        }
-    }
-
-    fn emit_style_changed(&mut self, applied_preset: Option<CandlePreset>) {
-        let candles = &self.style.candles;
-        let preset = applied_preset
-            .or_else(|| CandlePreset::detect(candles))
-            .map_or("custom", CandlePreset::log_value);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "CANDLE_STYLE_CHANGED",
-            revision = self.style_revision,
-            preset,
-            body_mode = ?candles.body_mode,
-            fill_opacity = candles.fill_opacity,
-            outline_opacity = candles.outline_opacity,
-            outline_width_px = candles.outline_width,
-            body_width_fraction = candles.body_width_frac,
-            wick_mode = ?candles.wick_color_mode,
-            wick_width_px = candles.wick_width,
-            chart_background_enabled = self.style.canvas.background_enabled,
-            chart_grid_enabled = self.style.canvas.grid_enabled,
-            action = "redraw_only",
-            "candle appearance changed"
-        );
-    }
-
-    /// Hot reload: about once a second, compare each file-backed script's
-    /// mtime; a changed file is re-read and sent as a Reload — recompiled
-    /// and replayed on success, or flagged stale (the last good version
-    /// keeps running) on errors. The mtime updates even when the compile
-    /// fails, so a broken save does not re-fire every second.
-    fn poll_script_files(&mut self) {
-        if self.script_files.is_empty()
-            || self.last_script_poll.elapsed() < SCRIPT_RELOAD_POLL_INTERVAL
-        {
-            return;
-        }
-        self.last_script_poll = Instant::now();
-        let mut reloads: Vec<(TabSlot, String, String)> = Vec::new();
-        for (owner, index, seen_mtime) in &mut self.script_files {
-            let Some((path, mtime)) = self.script_library.file_info(*index) else {
-                continue;
-            };
-            if mtime == *seen_mtime {
-                continue;
-            }
-            *seen_mtime = mtime;
-            match std::fs::read_to_string(&path) {
-                Ok(text) => {
-                    let name = path
-                        .file_name()
-                        .map(|n| n.to_string_lossy().into_owned())
-                        .unwrap_or_else(|| path.display().to_string());
-                    reloads.push((*owner, name, text));
-                }
-                Err(error) => tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "INDICATOR_SCRIPT_UNREADABLE",
-                    script = %path.display(),
-                    error = %error,
-                    action = "reload_skipped",
-                    "cannot re-read a changed indicator script"
-                ),
-            }
-        }
-        for (owner, name, text) in reloads {
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "INDICATOR_SCRIPT_RELOAD",
-                script = %name,
-                tab = owner.tab,
-                pane = ?owner.side,
-                action = "recompile_and_replay",
-                "indicator script changed on disk"
-            );
-            // To the worker that owns the slot: the same script loaded on two
-            // panes is two slots, and a Reload sent to the wrong one addresses
-            // whatever indicator happens to share its number there.
-            if let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == owner.tab) {
-                tab.pane_mut(owner.side)
-                    .indicator_worker
-                    .send(IndicatorCommand::Reload {
-                        slot: owner.slot,
-                        source: IndicatorSource::Script { name, text },
-                    });
-            }
-        }
-    }
-
-    /// Add one of the built-in indicators to the focused pane and register how
-    /// it restores.
-    ///
-    /// The kind travels to the worker as-is: an id no build ships becomes an
-    /// error slot naming it, which is why there is no fallback arm here. The
-    /// one this replaced turned every unrecognised kind into an EMA, so a
-    /// mistyped id put an indicator on the chart that nobody had asked for.
-    fn add_native_indicator(&mut self, id: &str) -> SlotId {
-        let kind = SavedKind::native(id);
-        let source = IndicatorSource::Native {
-            id: id.to_owned(),
-            values: Vec::new(),
-        };
-        let slot = self.focused_pane_mut().add_indicator(source);
-        let owner = self.target_slot(slot);
-        self.slot_kinds.push((owner, kind.clone()));
-        // Every other pane of every tab gets the same indicator now; the
-        // settled reconciliation binds the layout's copy of it.
-        self.mirror_add(owner, &kind);
-        self.note_indicator_edit_at(owner.tab, owner.side);
-        slot
-    }
-
-    /// An indicator edit happened on the focused pane. Edits that know their
-    /// pane call [`Self::note_indicator_edit_at`] directly.
-    fn mark_indicator_state_dirty(&mut self) {
-        let (tab, side) = {
-            let tab = self.active_tab();
-            (tab.id, tab.focused_side())
-        };
-        self.note_indicator_edit_at(tab, side);
-    }
-
-    /// Undo the dirty mark an add just set — for indicators an env var asked
-    /// for rather than the user. The slot still exists and still works; it
-    /// simply does not enter the persisted set.
-    fn forget_last_indicator_state_change(&mut self) {
-        self.slot_kinds.pop();
-    }
-
-    /// Apply layout-placed hide flags and styles once their views exist.
-    ///
-    /// The layout itself is written by the edit that changed it — see
-    /// `layout_wiring` — so nothing here reads a view back.
-    fn maintain_indicator_state(&mut self) {
-        self.apply_pending_indicator_state();
-    }
-
-    /// What a pane's layer menu could not switch itself.
-    ///
-    /// Drained right after the canvas, so the frame that clicked the entry is
-    /// the frame that applies it. Both wishes reach the real owner — the shared
-    /// style, the indicator state file — instead of a second copy on the pane.
-    fn apply_layer_actions(&mut self) {
-        let actions = std::mem::take(&mut self.layer_actions);
-        if let Some(visible) = actions.grid {
-            self.style.canvas.grid_enabled = visible;
-            // The appearance panel's own edits bump this; the renderer and the
-            // style log both read it to know something moved.
-            self.style_revision = self.style_revision.saturating_add(1);
-        }
-        if actions.indicators_changed {
-            self.mark_indicator_state_dirty();
-        }
-        if actions.footprint_changed {
-            crate::footprint_config::save(
-                self.workspace.footprint_settings_path(),
-                &self.footprint_config,
-            );
-        }
-        if actions.open_footprint_settings {
-            self.surfaces.footprint_settings.open();
-        }
-    }
-
-    /// Settle every tab's paper panel and hand its acknowledgement to the
-    /// window's one toast.
-    ///
-    /// # One lane, and what that cost
-    ///
-    /// The panel used to draw its own toast: the same `CENTER_BOTTOM` anchor
-    /// as `ToastSurface`, 96px up instead of 44, on a 4-second clock instead
-    /// of 8. Two acknowledgements could therefore sit in one lane, at two
-    /// heights, disagreeing about how long an acknowledgement lasts. There is
-    /// one now, and this is where the panel's messages join it.
-    ///
-    /// # Every tab, not just the one on screen
-    ///
-    /// `settle` runs for all of them because the jobs it finishes — an
-    /// export, an import — belong to the tab that started them, and a trader
-    /// who starts an export and then looks at another chart should not have
-    /// to come back for it to land. The acknowledgements follow: a stop
-    /// filling on a chart the trader is not looking at is precisely the news
-    /// they most need, and dropping it silently is what the old per-tab toast
-    /// did.
-    ///
-    /// A message from a background tab is **named**, because an unlabelled
-    /// "SIM: dropped at the fill" would read as being about the chart on
-    /// screen.
-    ///
-    /// # Which message wins a slot that holds one
-    ///
-    /// The watched tab's, always: it carries no prefix and is posted last, so
-    /// it takes the slot from any background message raised on the same
-    /// frame. Among background tabs the **first** in tab order wins and the
-    /// rest of that frame are dropped — the same first-wins rule
-    /// `SurfaceResponse::merge` uses for a request that carries a value, so
-    /// the window has one tie-break rule rather than two. Posting each of
-    /// them in turn would look like it showed them all and would in fact
-    /// show whichever `tabs.iter()` reached last, which is tab order deciding
-    /// in silence.
-    fn settle_paper_panels(&mut self, now: Instant) {
-        let Self {
-            tabs,
-            active_tab,
-            surfaces,
-            ..
-        } = self;
-        let mut watched = None;
-        let mut background = None;
-        for (index, tab) in tabs.iter_mut().enumerate() {
-            tab.paper.settle();
-            let Some(message) = tab.paper.take_toast() else {
-                continue;
-            };
-            if index == *active_tab {
-                watched = Some(message);
-            } else if background.is_none() {
-                // The interpunct is the window's own separator — the status
-                // bar, the tape's axis caption and the layout strip all use
-                // it, and the messages themselves already carry a colon
-                // (`SIM: …`). A second one would read as two labels.
-                background = Some(format!("{} · {message}", tab.symbol));
-            }
-        }
-        if let Some(message) = background {
-            surfaces.toast.note(message, now);
-        }
-        if let Some(message) = watched {
-            surfaces.toast.note(message, now);
-        }
-    }
-
-    /// Apply what the footprint settings window settled on.
-    ///
-    /// Whatever is edited also becomes the window default, which is what a
-    /// trader configuring their first chart means; a second chart diverges
-    /// only when they configure it too.
-    fn apply_footprint_change(&mut self, change: crate::surfaces::FootprintChange) {
-        let side = self.active_tab().focused_side();
-        match change {
-            crate::surfaces::FootprintChange::Applied(edited) => {
-                self.active_tab_mut().pane_mut(side).footprint_override = Some((*edited).clone());
-                self.footprint_config = *edited;
-                crate::footprint_config::save(
-                    self.workspace.footprint_settings_path(),
-                    &self.footprint_config,
-                );
-            }
-            crate::surfaces::FootprintChange::ResetToDefault => {
-                self.active_tab_mut().pane_mut(side).footprint_override = None;
-            }
-        }
-    }
-
-    /// The visibility this app persists, as one bit per layer.
-    ///
-    /// Read off the active tab's flow pane: the file records the canvas
-    /// quantick is built around, the same scope the indicator state file has
-    /// (see [`Self::maintain_indicator_state`]). A tab's second pane opens
-    /// matching it and is in-session from there.
-    fn layer_mask(&self) -> u32 {
-        self.active_tab().flow_pane.layer_mask(&self.style)
-    }
-
-    /// Save the layer visibility when it differs from what is on disk.
-    ///
-    /// Called once per frame instead of from each switch: the layers are owned
-    /// by four different pieces of chrome, and a save hook on each is four
-    /// chances to forget one.
-    fn maintain_chart_layers(&mut self) {
-        let mask = self.layer_mask();
-        // Layer state is per-pane, and this reads the *active* tab's flow pane
-        // — so activating a tab whose chart is set up differently changes the
-        // mask with nobody having touched a switch. Left alone, Ctrl+Tab
-        // records the other tab's opinion as the trader's choice, thrashes the
-        // file between two of them, and fills the log below with switches no
-        // hand moved. A different chart answering is a re-baseline, not an
-        // edit; the file keeps whatever the last real switch put there.
-        let tab = self.active_tab().id;
-        if tab != self.workspace.layers().tab() {
-            self.workspace.layers_mut().rebaseline(tab, mask);
-            return;
-        }
-        if mask == self.workspace.layers().mask() {
-            return;
-        }
-        // Name every switch that moved, before writing it down.
-        //
-        // This file is the trader's own answer, and it outranks the shipped
-        // default from the next launch on — so a layer that goes off without a
-        // click is not a display glitch, it is a choice attributed to someone
-        // who never made it, and it lasts. The bug that motivated this line was
-        // exactly that shape and cost a day to chase, because the only record
-        // was the file itself: a trio of `false`s with no timestamp, no
-        // sequence and nothing to say whether a hand had been near them.
-        //
-        // Off the frame path in every sense that matters: the mask compare
-        // above already gates it, so this runs on the frames a switch actually
-        // moves — a handful in a session — and not one of the other 60 a
-        // second.
-        let flipped = mask ^ self.workspace.layers().mask();
-        for (bit, layer) in ChartLayer::ALL.into_iter().enumerate() {
-            if flipped & (1 << bit) == 0 {
-                continue;
-            }
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "CHART_LAYER_SWITCHED",
-                layer = layer.id(),
-                on = mask & (1 << bit) != 0,
-                action = "persist_switch",
-                "a chart layer switch moved; recording it as the trader's choice"
-            );
-        }
-        chart_layers::save(
-            self.workspace.chart_layers_path(),
-            &self.active_tab().flow_pane.layer_states(&self.style),
-        );
-        self.workspace.layers_mut().record(mask);
-    }
-
-    /// Apply the saved layer visibility to the tab the app opened with.
-    ///
-    /// Runs before the autostart env vars so an explicit `QUANTICK_*_AUTOSTART`
-    /// still wins for the run it was set on: a validation session asks for the
-    /// heatmap on the command line and gets it, whatever the file remembers.
-    fn restore_chart_layers(&mut self) {
-        let defaults = chart_layers::load(self.workspace.chart_layers_path());
-        // Whatever the file said (including nothing at all) is now on screen;
-        // only a change from here is worth another write.
-        if defaults.is_empty() {
-            // Only reachable when the *shipped* config failed to parse, since
-            // every other path in `load` falls back to it — a build-time
-            // mistake, and the one launch where saying nothing would be worst:
-            // the chart opens on whatever the code decided and no one is told
-            // the product's own answer never arrived.
-            tracing::error!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "CHART_LAYERS_UNAVAILABLE",
-                path = %self.workspace.chart_layers_path().display(),
-                action = "keep_code_defaults",
-                "no layer visibility to apply; the shipped config did not parse"
-            );
-            let (tab, mask) = (self.active_tab().id, self.layer_mask());
-            self.workspace.layers_mut().rebaseline(tab, mask);
-            return;
-        }
-        if let Some(grid) = defaults.get(&ChartLayer::Grid) {
-            self.style.canvas.grid_enabled = *grid;
-        }
-        self.apply_layer_defaults(&defaults);
-        let (tab, mask) = (self.active_tab().id, self.layer_mask());
-        self.workspace.layers_mut().rebaseline(tab, mask);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "CHART_LAYERS_RESTORED",
-            path = %self.workspace.chart_layers_path().display(),
-            // `off`, not `hidden`: the map now always speaks for every layer,
-            // the shipped-off `backfill_divider` included, so a count over it
-            // is no longer "how many the trader switched off". Renamed rather
-            // than quietly redefined — a field that keeps its name and changes
-            // its meaning misleads every dashboard already reading it.
-            off = defaults.values().filter(|visible| !**visible).count(),
-            layers = defaults.len(),
-            "chart layer visibility restored"
-        );
-    }
-
-    /// Put every open pane on the saved visibility, once, at startup.
-    ///
-    /// A tab opened later does *not* come through here: it inherits the layers
-    /// the active tab is showing at that moment (`inherited_layers` in
-    /// [`Self::adopt_tab`]), which is the live state rather than the map read
-    /// off disk during boot. The two policies differ on purpose — see the note
-    /// there — so a reader arriving here for new-tab behaviour is in the wrong
-    /// function.
-    fn apply_layer_defaults(&mut self, states: &std::collections::BTreeMap<ChartLayer, bool>) {
-        for tab in &mut self.tabs {
-            // Every pane, by address: a default applied to "the flow pane and
-            // the time pane" left the second stacked chart on whatever the
-            // previous defaults were, so one canvas drew the same layer two
-            // ways.
-            for pane in tab.panes_mut() {
-                pane.apply_layer_states(states);
-            }
-        }
-    }
-
-    /// The window as it stands, in the form the workspace file records.
-    ///
-    /// Read off the live state rather than accumulated as it changes: the
-    /// arrangement is a dozen fields spread over the tabs and the chrome, and
-    /// a second copy maintained by every control that moves one of them would
-    /// be a dozen chances to forget. Saving is rare and event-driven, so
-    /// reading them all at once costs nothing anyone can see.
-    fn capture_workspace(&self) -> ui_state::Workspace {
-        let (tabs, chrome) = self.capture_arrangement();
-        ui_state::Workspace::new(
-            self.workspace.session().save_on_exit(),
-            self.window_size,
-            self.active_tab,
-            tabs,
-            Some(chrome),
-        )
-        // Every write rewrites the whole file, so the bookmarks have to ride
-        // along or saving the startup screen would silently delete them.
-        .with_saved(self.workspace.session().bookmarks().to_vec())
-        // And the recent workspace files, for the same reason: a save that
-        // dropped them would empty the Open-recent menu every time the
-        // trader saved their layout.
-        .with_recent(self.workspace.session().recent().to_vec())
-        // And so does the replay folder, for exactly the same reason: a save
-        // that dropped it would send the browser back to nowhere on the next
-        // launch, which is the failure this field was added to end. The
-        // trader's *pick*, never the folder in use — a run under
-        // `QUANTICK_REPLAY_DIR` must not write a QA scratch path into their
-        // workspace, and accepting the default home is not a choice either.
-        .with_replay_folder(self.replay_view.stored_pick().map(str::to_owned))
-        .with_replay_day_before(self.replay_view.stored_day_before())
-        // And the starred tools, for the third time and the same reason. They
-        // are already on disk the moment the star is clicked; riding along
-        // here keeps a full-file write from erasing what the star wrote.
-        .with_favorites(self.starred_tool_ids())
-    }
-
-    /// The rail's pinned section as tool ids, in star order — the form the
-    /// workspace file keeps it in.
-    fn starred_tool_ids(&self) -> Vec<String> {
-        self.toolrail
-            .favorites()
-            .iter()
-            .map(|tool| tool.id().to_owned())
-            .collect()
-    }
-
-    /// Put the chrome back — the mirror of the [`Self::capture_arrangement`]
-    /// half that produced it.
-    ///
-    /// One function because there are two callers and no way for the compiler
-    /// to notice when only one of them learns a new field: the startup
-    /// workspace and a named bookmark describe the same thing, and
-    /// [`ui_state::NamedArrangement`] says so in as many words. Restoring them
-    /// through two copies of the same eight lines is how a field comes to
-    /// persist but never come back from a bookmark — a bug with no compile
-    /// error behind it.
-    ///
-    /// The starred tools are the one thing it does not speak for: they live at
-    /// the file's own level rather than inside an arrangement
-    /// ([`ui_state::Workspace::favorite_tools`]), precisely so that opening a
-    /// bookmark cannot rearrange the rail the trader curated.
-    fn restore_chrome(&mut self, chrome: &ui_state::SavedChrome) {
-        self.tz = TzOffset::new(chrome.timezone_minutes);
-        self.dock
-            .restore(chrome.dock_visible, chrome.dock_tab.map(Into::into));
-        self.toolrail.set_dock(chrome.rail_dock.into());
-        self.toolrail.set_visible(chrome.rail_visible);
-        self.show_perf = chrome.perf_readings;
-        self.progressive_history = chrome.progressive_history;
-        // A token this release does not know keeps the reach it had — the
-        // default on startup, whatever the trader picked when a bookmark is
-        // opened mid-session. Never a silent fallback to something else: the
-        // reach decides how much a press fetches.
-        if let Some(reach) = chrome
-            .history_reach
-            .as_deref()
-            .and_then(history_reach::HistoryReach::from_token)
-        {
-            self.history_reach = reach;
-        }
-        // Through the setter, so a hand-edited workspace cannot restore a span
-        // the campaign could never reach.
-        if let Some(minutes) = chrome.history_reach_span_minutes {
-            self.set_history_reach_span_minutes(minutes);
-        }
-        self.venue_lead_in = chrome.venue_lead_in;
-        self.surfaces
-            .drawing_chrome
-            .restore_inspector_position(chrome.inspector_position);
-    }
-
-    /// The tabs and the chrome as they stand — the part a startup workspace
-    /// and a named one describe identically, so both capture through here.
-    fn capture_arrangement(&self) -> (Vec<ui_state::SavedTab>, ui_state::SavedChrome) {
-        let tabs = self
-            .tabs
-            .iter()
-            .map(|tab| ui_state::SavedTab {
-                feed: tab.feed_id.clone(),
-                symbol: tab.symbol.clone(),
-                layout: tab.layout.into(),
-                split_fraction: Some(tab.split_fraction),
-                context_collapsed: tab.context_collapsed,
-                focus: Some(ui_state::SavedFocus::from_side(tab.focused_side()).0),
-                focus_slot: ui_state::SavedFocus::from_side(tab.focused_side()).1,
-                flow_bars: tab.flow_pane.state.spec().to_config_string(),
-                // Only a pane that exists has an interval worth recording; a
-                // tab that never showed the split restores on the default,
-                // which is what it had.
-                time_bars: tab
-                    .time_pane()
-                    .map(|pane| pane.state.spec().to_config_string()),
-                context_bars: tab
-                    .time_panes
-                    .iter()
-                    .map(|pane| pane.state.spec().to_config_string())
-                    .collect(),
-                flow_layout: tab.flow_pane.layout.map(|layout| layout.0),
-                context_layouts: tab
-                    .time_panes
-                    .iter()
-                    .map(|pane| {
-                        pane.layout
-                            .map_or(crate::ui_state::LAYOUT_UNRECORDED, |layout| layout.0)
-                    })
-                    .collect(),
-                flow_legend_collapsed: tab.flow_pane.legend_collapsed,
-                // A tab with no time pane has no second legend, and `false`
-                // is what it will restore into when one is opened: a pane
-                // that never existed cannot have been folded.
-                time_legend_collapsed: tab.time_pane().is_some_and(|pane| pane.legend_collapsed),
-            })
-            .collect();
-        let chrome = ui_state::SavedChrome {
-            timezone_minutes: self.tz.minutes(),
-            dock_visible: self.dock.visible(),
-            dock_tab: self.dock.tab().map(Into::into),
-            rail_visible: self.toolrail.visible(),
-            rail_dock: self.toolrail.dock().into(),
-            perf_readings: self.show_perf,
-            // Never written any more: the stars are a standing choice and live
-            // at the top of the file. An arrangement that carried a copy would
-            // be an arrangement that could overwrite them on open.
-            legacy_favorite_tools: Vec::new(),
-            progressive_history: self.progressive_history,
-            // The default writes no key: a workspace that says nothing about
-            // the reach restores the press the button has always had, which is
-            // exactly what the default is.
-            // Written whenever it differs from what the config seeds, so a
-            // workspace only carries an opinion its owner actually formed.
-            history_reach_span_minutes: (self.history_reach_span_minutes
-                != self.config.history.reach_span_minutes)
-                .then_some(self.history_reach_span_minutes),
-            history_reach: (self.history_reach != history_reach::HistoryReach::default())
-                .then(|| self.history_reach.token().to_owned()),
-            venue_lead_in: self.venue_lead_in,
-            inspector_position: self.surfaces.drawing_chrome.remembered_inspector_position(),
-        };
-        (tabs, chrome)
-    }
-
-    /// Ask the operating system where to put a workspace file, off the UI
-    /// thread. One dialog at a time, the trades-folder picker's own pattern.
-    ///
-    /// The cockpit is written to its stores *first*, so what the file gets is
-    /// the screen as it stands rather than whatever was last flushed —
-    /// indicator state is written debounced, and an export that raced it
-    /// would quietly save a cockpit the trader never had.
-    fn open_workspace_export_picker(&mut self) {
-        if self.workspace.picker_open() {
-            return;
-        }
-        let start = crate::workspace_bundle::default_dir();
-        let suggested = crate::workspace_bundle::file_name_for(&self.suggested_workspace_name());
-        let (sender, receiver) = std::sync::mpsc::channel();
-        std::thread::Builder::new()
-            .name("quantick-workspace-export-picker".into())
-            .spawn(move || {
-                let mut dialog = rfd::FileDialog::new()
-                    .set_title("Export workspace")
-                    .add_filter("quantick workspace", &["toml"])
-                    .set_file_name(suggested);
-                if let Some(start) = start {
-                    let _ = std::fs::create_dir_all(&start);
-                    dialog = dialog.set_directory(&start);
-                }
-                let _ = sender.send(dialog.save_file());
-            })
-            .expect("spawn workspace export picker thread");
-        self.workspace.open_picker(WorkspacePick::Export, receiver);
-    }
-
-    /// The same, for choosing a workspace file to open.
-    fn open_workspace_import_picker(&mut self) {
-        if self.workspace.picker_open() {
-            return;
-        }
-        let start = crate::workspace_bundle::default_dir();
-        let (sender, receiver) = std::sync::mpsc::channel();
-        std::thread::Builder::new()
-            .name("quantick-workspace-import-picker".into())
-            .spawn(move || {
-                let mut dialog = rfd::FileDialog::new()
-                    .set_title("Open workspace")
-                    .add_filter("quantick workspace", &["toml"]);
-                if let Some(start) = start
-                    && start.is_dir()
-                {
-                    dialog = dialog.set_directory(&start);
-                }
-                let _ = sender.send(dialog.pick_file());
-            })
-            .expect("spawn workspace import picker thread");
-        self.workspace.open_picker(WorkspacePick::Import, receiver);
-    }
-
-    /// Land whatever the dialog answered.
-    fn poll_workspace_picker(&mut self) {
-        let Some((intent, receiver)) = self.workspace.picker() else {
-            return;
-        };
-        let choice = match receiver.try_recv() {
-            Ok(choice) => choice,
-            Err(std::sync::mpsc::TryRecvError::Empty) => return,
-            // The dialog thread died without answering — no display server,
-            // a COM failure. Treated as "no answer yet" this would leave the
-            // field set forever and both menu entries silently dead, since
-            // each refuses to open a second dialog.
-            Err(std::sync::mpsc::TryRecvError::Disconnected) => {
-                self.workspace.close_picker();
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "WORKSPACE_PICKER_LOST",
-                    action = "picker_reset",
-                    "the file dialog closed without an answer"
-                );
-                self.note_workspace(
-                    "The file chooser could not open — see the log. Try again.".to_owned(),
-                );
-                return;
-            }
-        };
-        let intent = *intent;
-        self.workspace.close_picker();
-        // A cancelled dialog is an answer, not a failure: say nothing.
-        let Some(path) = choice else { return };
-        match intent {
-            WorkspacePick::Export => self.export_workspace_to(&path),
-            WorkspacePick::Import => self.import_workspace_from(&path),
-        }
-    }
-
-    /// A starting name for an export: what the cockpit is showing, so the
-    /// trader does not have to invent one to get a usable file name.
-    fn suggested_workspace_name(&self) -> String {
-        let tab = self.active_tab();
-        format!(
-            "{} {}",
-            tab.symbol,
-            tab.flow_pane.state.spec().to_config_string()
-        )
-    }
-
-    /// Write every store that is still only in memory, so a bundle captured
-    /// next describes the screen rather than the last flush.
-    fn flush_cockpit_stores(&mut self) {
-        // The layouts file is written debounced, off the frame path. An
-        // export is the one moment worth paying it immediately, or the
-        // bundle would carry the layouts as they stood a second ago.
-        self.flush_layouts();
-        self.maintain_chart_layers();
-    }
-
-    /// Export the whole cockpit to one file, and say what happened.
-    fn export_workspace_to(&mut self, path: &std::path::Path) {
-        // Here rather than before the dialog: these two writes are how the
-        // bundle comes to describe the screen instead of the last flush, and
-        // doing them up front meant a trader who pressed Cancel had still
-        // silently redefined what the app opens on. It also keeps the harness
-        // hook on exactly the menu's path.
-        self.save_workspace("export");
-        self.flush_cockpit_stores();
-        let name = crate::workspace_bundle::recent_label(path);
-        let outcome = crate::workspace_bundle::capture(
-            &name,
-            crate::store_home::COCKPIT_STORES,
-            &crate::workspace_bundle::live_paths,
-        )
-        .and_then(|bundle| crate::workspace_bundle::write(path, &bundle).map(|()| bundle.len()));
-        match outcome {
-            Ok(stores) => {
-                crate::workspace_bundle::remember_recent(
-                    self.workspace.session_mut().recent_mut(),
-                    path,
-                );
-                self.refresh_recent_workspaces();
-                self.save_workspace("export_recent");
-                tracing::info!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "WORKSPACE_EXPORTED",
-                    path = %path.display(),
-                    stores,
-                    action = "workspace_written",
-                    "workspace exported"
-                );
-                self.note_workspace(format!(
-                    "Workspace exported to {} — {stores} settings groups",
-                    path.display()
-                ));
-            }
-            Err(error) => {
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "WORKSPACE_EXPORT_FAILED",
-                    path = %path.display(),
-                    %error,
-                    action = "workspace_not_written",
-                    "could not export the workspace"
-                );
-                self.note_workspace(format!("Workspace not exported — {error}"));
-            }
-        }
-    }
-
-    /// Open a workspace file over the live cockpit.
-    ///
-    /// Refused whole or applied whole: [`crate::workspace_bundle::apply`] checks
-    /// every section before writing any, so a bad file leaves the screen
-    /// exactly as it was and says why. What reaches the disk is then read
-    /// back into the running app, because a cockpit that only changed on disk
-    /// would be overwritten by this session's own save on exit.
-    fn import_workspace_from(&mut self, path: &std::path::Path) {
-        // What the debounce still holds is written first, or a bundle with no
-        // layouts section would reload a file a second behind the screen.
-        self.flush_layouts();
-        let outcome = crate::workspace_bundle::read(path).and_then(|bundle| {
-            crate::workspace_bundle::apply(
-                &bundle,
-                crate::store_home::COCKPIT_STORES,
-                &crate::workspace_bundle::live_paths,
-            )
-        });
-        match outcome {
-            Ok(written) => {
-                let stores = written.len();
-                self.reload_cockpit_stores(&written);
-                crate::workspace_bundle::remember_recent(
-                    self.workspace.session_mut().recent_mut(),
-                    path,
-                );
-                self.refresh_recent_workspaces();
-                // The recent list lives in the workspace file the import just
-                // replaced, so it has to be written back after the reload —
-                // otherwise opening a file would forget that it was opened.
-                self.save_workspace("import_recent");
-                tracing::info!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "WORKSPACE_IMPORTED",
-                    path = %path.display(),
-                    stores,
-                    action = "cockpit_replaced",
-                    "workspace imported"
-                );
-                self.note_workspace(format!(
-                    "Workspace \"{}\" opened — {stores} settings groups restored",
-                    crate::workspace_bundle::recent_label(path)
-                ));
-            }
-            Err(error) => {
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "WORKSPACE_IMPORT_REFUSED",
-                    path = %path.display(),
-                    %error,
-                    action = "cockpit_unchanged",
-                    "workspace refused"
-                );
-                self.note_workspace(format!("Workspace not opened — {error}"));
-            }
-        }
-    }
-
-    /// Read every cockpit store back into the running app.
-    ///
-    /// Called after an import wrote them. Each store goes through the loader
-    /// it already had, so an imported cockpit is restored by exactly the code
-    /// that restores one at startup — a second, import-only restore path is
-    /// how the two would drift.
-    fn reload_cockpit_stores(&mut self, imported: &[&str]) {
-        self.added_symbols = symbols_file::load(self.workspace.symbols_path());
-        self.drawing_presets = drawings::presets::PresetStore::load_from(
-            drawings::presets::PresetStore::default_path(),
-        );
-        self.footprint_config =
-            crate::footprint_config::load(self.workspace.footprint_settings_path());
-        self.surfaces.footprint_settings.reload_presets();
-        self.indicator_presets =
-            preset_file::PresetStore::load(self.workspace.indicator_presets_path());
-
-        // The tab strip first, and *before* the indicators: the restore adds
-        // each indicator to whatever pane is focused right now, so the tabs
-        // have to be the imported ones — and the focus on the pane the file
-        // describes — before a single indicator is added. Getting this order
-        // wrong puts a trader's imported indicators on the tab they happened
-        // to be looking at, or on its time pane.
-        let workspace =
-            ui_state::load(self.workspace.ui_state_path()).restore(&self.config.clone());
-        self.restore_workspace(workspace);
-        self.restore_chart_layers();
-
-        // The layouts come last, once the tabs are the imported ones: every
-        // pane is stripped and re-seeded from the imported file.
-        self.reload_layouts(imported);
-    }
-
-    /// Work out which remembered workspace files are still there.
-    ///
-    /// Called when the list changes, never from the menu body — see
-    /// [`Self::recent_on_disk`]. The stored list keeps every entry: a file on
-    /// a drive that is merely unplugged today comes back when it is plugged
-    /// in, and only the menu is filtered.
-    fn refresh_recent_workspaces(&mut self) {
-        let existing = crate::workspace_bundle::existing_recent(self.workspace.session().recent());
-        self.workspace.session_mut().set_recent_on_disk(existing);
-    }
-
-    /// Take every indicator off every pane.
-    ///
-    /// Straight off each pane's own collection rather than by walking
-    /// `slot_kinds`: that list is bookkeeping for the state *file*, and an
-    /// indicator can be on a pane without being in it — the autostart hooks
-    /// add without registering, and `forget_last_indicator_state_change` pops
-    /// an entry while leaving the indicator on screen. Clearing the list
-    /// would have left those behind for the imported set to stack on top of.
-    fn clear_indicators(&mut self) {
-        /// Empty one pane, view and worker alike.
-        fn strip(pane: &mut crate::pane::ChartPane) {
-            let slots: Vec<SlotId> = pane.indicators.all().iter().map(|view| view.slot).collect();
-            for slot in slots {
-                pane.indicators.remove(slot);
-                pane.indicator_worker.send(IndicatorCommand::Remove(slot));
-            }
-        }
-        for tab in &mut self.tabs {
-            // Every pane the tab holds, not the two it used to. `panes_mut`
-            // rather than `pane_mut(Time)`: the latter falls back to the flow
-            // pane when a tab was never split, which would strip it twice, and
-            // it stops at the *first* context chart — so the second stacked
-            // chart kept its indicators while `slot_kinds` was cleared out from
-            // under them, and the imported set stacked on top.
-            for pane in tab.panes_mut() {
-                strip(pane);
-            }
-        }
-        self.slot_kinds.clear();
-        self.operator_slots.clear();
-        self.script_files.clear();
-        self.pending_hidden.clear();
-        self.pending_styles.clear();
-        self.mark_indicator_state_dirty();
-    }
-
-    /// Show the trader where the cockpit is kept, and open it.
-    ///
-    /// The answer to "where does this thing save my setup?" — which, before
-    /// the durable home, had no single answer at all.
-    fn reveal_cockpit_home(&mut self) {
-        let Some(home) = crate::store_home::home() else {
-            self.note_workspace(
-                "This system reports no documents folder — quantick keeps the cockpit beside \
-                 wherever it was launched from"
-                    .to_owned(),
-            );
-            return;
-        };
-        self.note_workspace(format!("Cockpit saved in {}", home.display()));
-        // The journal panel's opener, not a second copy of the platform
-        // table: two of them means the next fix lands on one and misses the
-        // other. Best effort — the path is on the status line either way, so
-        // a system with no file manager still answers the question.
-        crate::paper_trading::reveal_folder(&home);
-    }
-
-    /// Write the workspace and say so on the status bar.
-    ///
-    /// The notice is the point of the explicit action: a trader who arranges a
-    /// cockpit and clicks Save wants to know it is kept, and "it looks the
-    /// same" is not an answer. A failed write says *that* instead — being told
-    /// "saved" and finding out at the next launch is the one outcome worth
-    /// engineering against.
-    fn save_workspace(&mut self, reason: &'static str) {
-        let workspace = self.capture_workspace();
-        let saved = ui_state::save(self.workspace.ui_state_path(), &workspace);
-        self.workspace.session_mut().note_write(saved);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "UI_STATE_SAVED",
-            path = %self.workspace.ui_state_path().display(),
-            tabs = workspace.tabs.len(),
-            saved,
-            reason,
-            action = if saved { "workspace_written" } else { "workspace_not_written" },
-            "workspace save"
-        );
-        self.note_workspace(if saved {
-            format!(
-                "Workspace saved — quantick opens on {} {}",
-                workspace.tabs.len(),
-                if workspace.tabs.len() == 1 {
-                    "chart tab"
-                } else {
-                    "chart tabs"
-                }
-            )
-        } else {
-            "Workspace could not be saved — see the log".to_owned()
-        });
-    }
-
-    /// Update *only* where the properties popup goes, in the workspace file as
-    /// it already stands. Says whether anything was written.
-    ///
-    /// Deliberately not [`Self::save_workspace`]. That one captures the whole
-    /// window — every tab, its market, its bar rule, the layout, the window
-    /// size — and adopts it as the startup screen. Parking a popup is not a
-    /// statement about any of that: a trader who opens six tabs to research
-    /// something, then nudges the popup out of the way, must not find those six
-    /// tabs waiting for them tomorrow. The switch that governs whole-window
-    /// saves says "when the window closes", and this write happens mid-session,
-    /// so it has no business speaking for the tabs.
-    ///
-    /// A file with no chrome section is left alone rather than created: there
-    /// is nothing to update, and inventing a startup workspace out of a drag
-    /// would undo a `Reset startup layout` the trader just asked for. The exit
-    /// save is what creates the file, and it carries the position with it.
-    ///
-    /// No toast either. The "Workspace saved" line answers a deliberate *Save*,
-    /// and repeating it after every small gesture would turn the window's one
-    /// acknowledgement channel into wallpaper — but the log still names the
-    /// write, so a position that went missing is answerable.
-    fn write_inspector_position(&mut self) -> bool {
-        let mut file = ui_state::load(self.workspace.ui_state_path());
-        let Some(chrome) = file.chrome.as_mut() else {
-            return false;
-        };
-        let position = self.surfaces.drawing_chrome.remembered_inspector_position();
-        if chrome.inspector_position == position {
-            return false;
-        }
-        chrome.inspector_position = position;
-        let saved = ui_state::save(self.workspace.ui_state_path(), &file);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "UI_STATE_POPUP_POSITION_SAVED",
-            path = %self.workspace.ui_state_path().display(),
-            parked = position.is_some(),
-            saved,
-            action = if saved { "position_written" } else { "position_not_written" },
-            "properties popup position"
-        );
-        saved
-    }
-
-    /// Write the bookmarks without disturbing the startup arrangement.
-    ///
-    /// Reads the file back and swaps only the named entries, rather than
-    /// capturing the live window: saving a bookmark must not redefine what the
-    /// app opens on, and `capture_workspace` describes the screen *now*, which
-    /// is exactly what the startup arrangement must not become.
-    fn write_bookmarks(&mut self) -> bool {
-        let bookmarks = self.workspace.session().bookmarks().to_vec();
-        let written = self.edit_workspace_file("UI_STATE_BOOKMARKS_WRITTEN", |file| {
-            file.saved = bookmarks;
-        });
-        self.workspace.session_mut().note_write(written);
-        written
-    }
-
-    /// Change one standing choice in the workspace file, leaving everything
-    /// else in it exactly as it was. `true` when the change reached the disk.
-    ///
-    /// This file holds three choices that are not descriptions of the screen —
-    /// the named bookmarks, the replay folder, the starred tools. Each is made
-    /// by a single click and each is written on the spot rather than at exit,
-    /// because "it forgot again" must not be one crash away. Each used to
-    /// hand-roll the same read-swap-write, and three copies were three chances
-    /// to differ: two carried `save_on_exit` through and one did not, so a
-    /// trader with autosave off who picked a replay folder before the file
-    /// existed had autosave quietly switched back on at the next launch.
-    ///
-    /// `save_on_exit` rides along here because it is the one live setting that
-    /// belongs to the *file* rather than to any arrangement inside it.
-    ///
-    /// A file this build cannot read is never rewritten — see
-    /// [`ui_state::load_for_edit`]. `workspace_saved` is deliberately not
-    /// touched: whether a startup *arrangement* exists is a different question
-    /// from whether this file does, and the caller answers it.
-    fn edit_workspace_file(
-        &mut self,
-        event_code: &'static str,
-        edit: impl FnOnce(&mut ui_state::Workspace),
-    ) -> bool {
-        let Some(mut file) = ui_state::load_for_edit(self.workspace.ui_state_path()) else {
-            self.note_workspace(
-                "The workspace file could not be read, so it was left alone — see the log"
-                    .to_owned(),
-            );
-            return false;
-        };
-        file.save_on_exit = self.workspace.session().save_on_exit();
-        edit(&mut file);
-        let written = ui_state::save(self.workspace.ui_state_path(), &file);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code,
-            path = %self.workspace.ui_state_path().display(),
-            written,
-            action = if written { "file_updated" } else { "file_not_written" },
-            "a standing choice was written to the workspace file"
-        );
-        written
-    }
-
-    /// Write down the replay folder the trader just pointed the browser at,
-    /// without disturbing anything else the workspace file holds.
-    ///
-    /// The same read-swap-write as [`Self::write_bookmarks`], and for the same
-    /// reason: this is a standing choice, not a description of the screen, so
-    /// it must not wait for a clean exit and must not drag the current
-    /// arrangement into the file with it.
-    fn write_replay_folder(&mut self, folder: Option<&str>) {
-        let stored = folder.map(str::to_owned);
-        let written = self.edit_workspace_file("REPLAY_FOLDER_REMEMBERED", |file| {
-            file.replay_folder = stored;
-        });
-        self.workspace.session_mut().note_write(written);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "REPLAY_FOLDER_REMEMBERED",
-            folder = folder.unwrap_or_default(),
-            written,
-            action = if folder.is_some() {
-                "store_replay_folder"
-            } else {
-                "forget_replay_folder"
-            },
-            "the replay folder is now the one this workspace opens on"
-        );
-    }
-
-    /// Write down the *day before* choice the trader just made, without
-    /// disturbing anything else the workspace file holds.
-    ///
-    /// The same read-swap-write as [`Self::write_replay_folder`], and for the
-    /// same reason: whether yesterday is on the chart is a standing choice,
-    /// not a description of the screen, so it must not wait for a clean exit.
-    fn write_replay_day_before(&mut self, enabled: bool) {
-        let written = self.edit_workspace_file("REPLAY_DAY_BEFORE_REMEMBERED", |file| {
-            file.replay_day_before = Some(enabled);
-        });
-        self.workspace.session_mut().note_write(written);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "REPLAY_DAY_BEFORE_REMEMBERED",
-            enabled,
-            written,
-            action = if enabled {
-                "join_day_before"
-            } else {
-                "chosen_day_only"
-            },
-            "the day before is now what this workspace opens recordings with"
-        );
-    }
-
-    /// Write down the tools the trader just starred or unstarred, without
-    /// disturbing anything else the workspace holds.
-    ///
-    /// The same read-swap-write as [`Self::write_replay_folder`], through the
-    /// same [`Self::edit_workspace_file`], with two things of its own.
-    ///
-    /// First, `save_on_exit` does not gate it. That switch governs whether
-    /// closing the window redefines the *arrangement* — which tabs open, how
-    /// the panes are split. A starred tool is not an arrangement, and a trader
-    /// who turned autosave off to stop their layout drifting has not asked to
-    /// rebuild their rail every session.
-    ///
-    /// Second, `workspace_saved` is left alone. It answers "is there a startup
-    /// arrangement to reset?", and starring a tool does not create one — a
-    /// fresh install whose only saved thing is a star would otherwise light up
-    /// a Reset entry that promises to forget a layout nobody ever saved.
-    fn write_favorites(&mut self) {
-        // A run under `QUANTICK_TOOL_FAVORITES` is wearing a rail the harness
-        // dressed it in, not one the trader curated, and the same guard the
-        // replay folder gets applies: a validation run must not write a QA
-        // list into the trader's workspace. The hook stages a screen; it does
-        // not make choices on their behalf.
-        if self.workspace.session().favorites_are_staged() {
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "TOOL_FAVORITES_NOT_WRITTEN",
-                action = "staged_by_hook",
-                "a run under QUANTICK_TOOL_FAVORITES does not write the rail down"
-            );
-            return;
-        }
-        let tools = self.starred_tool_ids();
-        let count = tools.len();
-        let written = self.edit_workspace_file("TOOL_FAVORITES_REMEMBERED", |file| {
-            file.favorite_tools = tools;
-        });
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "TOOL_FAVORITES_REMEMBERED",
-            tools = count,
-            written,
-            action = if written { "favorites_written" } else { "favorites_not_written" },
-            "the rail's pinned tools are now what this workspace opens on"
-        );
-    }
-
-    /// Keep the window as it stands under `name`.
-    ///
-    /// A bookmark, not a startup setting: what the app opens on is untouched.
-    /// The reason to name an arrangement is usually to have somewhere to come
-    /// back *to*, and a "save this so I can return to it" that also redefined
-    /// the opening screen would be the opposite of a safety net.
-    ///
-    /// An existing name is replaced rather than duplicated — that is what
-    /// "save as" means everywhere else, and it spares the menu a list of five
-    /// entries called "scalp".
-    fn save_named_workspace(&mut self, name: &str) {
-        let Some(name) = ui_state::clean_workspace_name(name) else {
-            self.note_workspace("A workspace needs a name".to_owned());
-            return;
-        };
-        let (tabs, chrome) = self.capture_arrangement();
-        let entry = ui_state::NamedArrangement {
-            name: name.clone(),
-            window: self.window_size,
-            active_tab: self.active_tab,
-            tabs,
-            chrome: Some(chrome),
-        };
-        let replaced = match self
-            .workspace
-            .session_mut()
-            .bookmarks_mut()
-            .iter_mut()
-            .find(|held| held.name == name)
-        {
-            Some(held) => {
-                *held = entry;
-                true
-            }
-            None => {
-                self.workspace.session_mut().bookmarks_mut().push(entry);
-                false
-            }
-        };
-        let written = self.write_bookmarks();
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "UI_STATE_NAMED_SAVED",
-            name = %name,
-            replaced,
-            saved = self.workspace.session().bookmarks().len(),
-            written,
-            action = if written { "bookmark_written" } else { "bookmark_not_written" },
-            "named workspace saved"
-        );
-        self.note_workspace(if written {
-            let verb = if replaced { "replaced" } else { "saved" };
-            format!("Workspace \"{name}\" {verb} — reopen it from Workspace → Open")
-        } else {
-            format!("\"{name}\" could not be saved — see the log")
-        });
-    }
-
-    /// Put the window back the way the bookmark called `name` recorded it.
-    ///
-    /// The saved markets are opened as new tabs and the tabs that were on
-    /// screen are closed afterwards, rather than the reverse: `close_tab`
-    /// refuses to close the last tab — a window with no market has nothing to
-    /// draw — so growing before shrinking is what lets the whole strip be
-    /// replaced. Closing goes through the same path a `Ctrl+W` takes, so a
-    /// simulated position ends in the labeled, journaled flatten the
-    /// paper-trading contract promises instead of vanishing with its tab.
-    ///
-    /// The startup workspace is left alone. Opening a bookmark is a thing you
-    /// do to *this session*; making it the opening screen is `Save workspace`,
-    /// one entry above.
-    fn open_named_workspace(&mut self, name: &str) {
-        let Some(entry) = self
-            .workspace
-            .session()
-            .bookmarks()
-            .iter()
-            .find(|held| held.name == name)
-            .cloned()
-        else {
-            self.note_workspace(format!("No workspace called \"{name}\""));
-            return;
-        };
-        if entry.tabs.is_empty() {
-            // `restore` drops empty bookmarks at load, so this is only
-            // reachable from a file edited under a running app.
-            self.note_workspace(format!("\"{name}\" has no market left to open"));
-            return;
-        }
-        let replaced = self.tabs.len();
-        for saved in &entry.tabs {
-            self.open_tab(
-                saved.feed.clone(),
-                saved.symbol.clone(),
-                BarSpec::parse(&saved.flow_bars).ok(),
-            );
-            let context_intervals =
-                saved_context_intervals(&saved.context_bars, saved.time_bars.as_deref());
-            let opened = self.tabs.len() - 1;
-            self.tabs[opened].restore_canvas(
-                CanvasLayout::from(saved.layout),
-                saved.split_fraction,
-                saved.context_collapsed,
-                saved.focus.map(|focus| focus.to_side(saved.focus_slot)),
-                &context_intervals,
-                LegendFold {
-                    flow: saved.flow_legend_collapsed,
-                    time: saved.time_legend_collapsed,
-                },
-            );
-            self.tabs[opened].set_opening_layouts(saved.flow_layout, &saved.context_layouts);
-        }
-        for _ in 0..replaced {
-            self.close_tab(0);
-        }
-        // The pinned section is deliberately untouched: the stars live beside
-        // the chrome rather than inside it, and [`Self::restore_chrome`] does
-        // not speak for them. A bookmark rearranges the cockpit; the tools the
-        // trader keeps at hand are not part of the arrangement, and a bookmark
-        // named before they starred anything used to wipe the rail on open.
-        if let Some(chrome) = &entry.chrome {
-            self.restore_chrome(chrome);
-        }
-        self.active_tab = entry.active_tab.min(self.tabs.len().saturating_sub(1));
-        let config = self.config.clone();
-        self.active_tab_mut().refresh_chip_label(&config);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "UI_STATE_NAMED_OPENED",
-            name = %name,
-            tabs = self.tabs.len(),
-            closed = replaced,
-            active = self.active_tab,
-            action = "replace_tab_strip",
-            "named workspace opened"
-        );
-        self.note_workspace(format!(
-            "Opened \"{name}\" — {} {}",
-            self.tabs.len(),
-            if self.tabs.len() == 1 {
-                "chart tab"
-            } else {
-                "chart tabs"
-            }
-        ));
-    }
-
-    /// Forget the bookmark called `name`. The window on screen is untouched —
-    /// deleting a bookmark throws away a way back, not the place you are.
-    fn delete_named_workspace(&mut self, name: &str) {
-        let before = self.workspace.session().bookmarks().len();
-        self.workspace
-            .session_mut()
-            .bookmarks_mut()
-            .retain(|held| held.name != name);
-        if self.workspace.session().bookmarks().len() == before {
-            return;
-        }
-        let written = self.write_bookmarks();
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "UI_STATE_NAMED_DELETED",
-            name = %name,
-            remaining = self.workspace.session().bookmarks().len(),
-            written,
-            action = if written { "bookmark_forgotten" } else { "file_not_written" },
-            "named workspace deleted"
-        );
-        self.note_workspace(if written {
-            format!("Workspace \"{name}\" deleted")
-        } else {
-            format!("\"{name}\" could not be deleted — see the log")
-        });
-    }
-
-    /// Forget the saved workspace: the next launch opens on the configured
-    /// defaults. The window on screen is deliberately left alone — a trader
-    /// resetting their *startup* layout mid-session has not asked to have the
-    /// charts they are reading rearranged under them.
-    fn forget_workspace(&mut self) {
-        // Reset clears the *startup* arrangement. The bookmarks survive it,
-        // because coming back after a reset is the whole reason to name one:
-        // deleting the safety net as part of the act it exists to undo would
-        // be the single worst thing this menu could do.
-        let bookmarks_kept = !self.workspace.session().bookmarks().is_empty();
-        // The starred tools survive it too, and for a plainer reason: they
-        // were never part of the arrangement being reset. Resetting a layout
-        // is not asking to rebuild the rail by hand — and the same goes for
-        // every other standing choice this file holds. The replay folder and
-        // the Open-recent list are facts about this installation; the entry
-        // resets a *layout* and must not quietly take them with it.
-        let stars = self.starred_tool_ids();
-        let kept = bookmarks_kept
-            || !stars.is_empty()
-            || !self.workspace.session().recent().is_empty()
-            || self.replay_view.stored_pick().is_some();
-        let bookmarks = self.workspace.session().bookmarks().to_vec();
-        let stars_kept = !stars.is_empty();
-        let forgotten = if kept {
-            // Edited rather than rebuilt from the defaults: writing a fresh
-            // `Workspace` would carry only what this function remembered to
-            // thread through it, and the fields it forgot would be reset by
-            // omission. Clearing the arrangement names what goes; everything
-            // unnamed stays by construction.
-            self.edit_workspace_file("UI_STATE_FORGOTTEN", |file| {
-                file.tabs.clear();
-                file.chrome = None;
-                file.window = None;
-                file.active_tab = 0;
-                file.saved = bookmarks;
-                file.favorite_tools = stars;
-            })
-        } else {
-            ui_state::forget(self.workspace.ui_state_path())
-        };
-        // The file still exists while it holds standing choices, so Reset
-        // stays available — it is now a no-op for the startup screen and the
-        // entry says as much. A reset that *failed* leaves the old
-        // arrangement on disk and so leaves the entry live: the trader has to
-        // be able to try again, and telling them "nothing saved yet" while the
-        // next launch still reopens the layout they discarded would be a lie.
-        self.workspace
-            .session_mut()
-            .set_saved(if forgotten { kept } else { true });
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "UI_STATE_FORGOTTEN",
-            path = %self.workspace.ui_state_path().display(),
-            forgotten,
-            bookmarks_kept = self.workspace.session().bookmarks().len(),
-            favorites_kept = self.toolrail.favorites().len(),
-            action = if forgotten { "open_on_config_defaults" } else { "workspace_kept" },
-            "workspace reset"
-        );
-        // What survived is named, never left to be discovered. A trader who
-        // resets a layout and is told nothing else assumes nothing else was
-        // kept — and would go looking for stars that are still there.
-        let survivors = match (bookmarks_kept, stars_kept) {
-            (true, true) => format!(
-                " {} saved {} and the starred tools kept.",
-                self.workspace.session().bookmarks().len(),
-                if self.workspace.session().bookmarks().len() == 1 {
-                    "workspace"
-                } else {
-                    "workspaces"
-                }
-            ),
-            (true, false) => format!(
-                " {} saved {} kept.",
-                self.workspace.session().bookmarks().len(),
-                if self.workspace.session().bookmarks().len() == 1 {
-                    "workspace"
-                } else {
-                    "workspaces"
-                }
-            ),
-            (false, true) => " The starred tools are kept.".to_owned(),
-            (false, false) => String::new(),
-        };
-        self.note_workspace(if forgotten {
-            format!(
-                "Startup layout reset — the next launch opens on the configured default.{survivors}"
-            )
-        } else {
-            "Workspace could not be reset — see the log".to_owned()
-        });
-    }
-
-    /// Hand the app the window it is drawing into.
-    ///
-    /// Called once from `main`, which is where eframe offers the handle: the
-    /// hook that needs it (`raw_input_hook`) is given only the input. One
-    /// registration line rather than a constructor argument, because every
-    /// other construction path — every test — wants the `None` this defaults
-    /// to, which is also what every non-Windows target gets.
-    pub fn attach_surface(&mut self, handle: &impl raw_window_handle::HasWindowHandle) {
-        self.surface = window_scale::SurfaceProbe::new(handle);
-    }
-
-    /// Take the window manager's own maximise, once, on the first frame.
-    ///
-    /// Through [`egui::ViewportCommand::Maximized`] rather than the viewport
-    /// builder's `with_maximized`: eframe 0.29 does not honour that flag beside
-    /// an `inner_size`, and a hook that silently opens a 1100×650 window while
-    /// reporting success is worse than no hook — a validation run would
-    /// photograph the wrong state and call it a pass. The command is the one
-    /// the platform runs when a hand hits the title bar, which is the state
-    /// this hook exists to reach.
-    fn apply_maximize_hook(&mut self, ctx: &egui::Context) {
-        if !self.harness.take_maximize() {
-            return;
-        }
-        ctx.send_viewport_cmd(egui::ViewportCommand::Maximized(true));
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "WINDOW_MAXIMIZE_AUTOSTART",
-            action = "maximize",
-            "QUANTICK_WINDOW_MAXIMIZED asked for the maximised layout"
-        );
-    }
-
-    /// Keep the window size the workspace would record, flush a popup the
-    /// trader just re-parked, and take the exit save when the window is
-    /// closing.
-    ///
-    /// **Per-frame cost**: two reads off the frame's own input state, a float
-    /// compare and a `bool` test. No save is on this path — each of the three
-    /// happens on one frame: the frame a drag ends, and the frame the close is
-    /// requested, when the window is going away anyway.
-    ///
-    /// The size is tracked here rather than read at exit because by then the
-    /// viewport has already been asked to close: what a workspace should
-    /// remember is the window the trader was working in, not whatever the
-    /// platform reports on the way out.
-    fn maintain_workspace(&mut self, ctx: &egui::Context) {
-        let (size, closing) = ctx.input(|input| {
-            let viewport = input.viewport();
-            (
-                viewport
-                    .inner_rect
-                    .map(|rect| [rect.width(), rect.height()]),
-                viewport.close_requested(),
-            )
-        });
-        if let Some(size) = size
-            && size[0] > 0.0
-            && size[1] > 0.0
-        {
-            self.window_size = Some(size);
-        }
-        // Where the trader parks the properties popup is kept the moment the
-        // hand comes off it, without a trip through the Workspace menu: the
-        // window they dragged out of the way is the window they expect back,
-        // and a memory that only survived a clean exit would lose it to the
-        // one session that ended badly. Only that one field is written — see
-        // [`Self::write_inspector_position`] for why a drag must not adopt the
-        // tab strip as a startup screen.
-        //
-        // Gated on the same switch the exit save is, because the menu already
-        // promises it: "Off, only Save workspace changes what quantick opens
-        // on." A trader who curates their startup layout by hand still gets the
-        // position within the session — it is live state — and an explicit Save
-        // still records it.
-        //
-        // The closing frame takes the exit save instead: it writes the whole
-        // window, this position included, so running both would serialise the
-        // same file twice on the way out.
-        if closing && self.workspace.session().save_on_exit() {
-            self.inspector_position_dirty = false;
-            self.save_workspace("exit");
-        } else if std::mem::take(&mut self.inspector_position_dirty)
-            && self.workspace.session().save_on_exit()
-        {
-            self.write_inspector_position();
-        }
-    }
-
-    /// Post a Workspace-menu answer through the window's one acknowledgement
-    /// channel ([`Toast`]).
-    ///
-    /// No Undo: the file it replaced is gone, and `Reset startup layout` is
-    /// the honest way back rather than a button that pretends otherwise.
-    fn note_workspace(&mut self, message: String) {
-        self.surfaces.toast.note(message, Instant::now());
     }
 }
