+++ crates/guards/size-baseline.txt
@@ -105,9 +105,24 @@
 # taken and the fall is a trade of nothing for 2,153. The two functions that
 # are the pane, `handle_navigation` and `draw_chart`, stayed put on purpose:
 # they are the next cut, and they need a shape rather than a new address.
-!budget 57394
+#
+# Then lowered by 2,160 more -- `app.rs` from 7,499 to 5,339 -- by `--tighten`
+# after the third cut of the same sprint moved three `impl QuantickApp` groups
+# to siblings under `crates/app/src/app/`: the workspace store's write half,
+# the indicator manager and the chart-layer wiring. Same trade as the two cuts
+# before it and the same reason it costs nothing here: `workspace_save.rs`
+# (1,124 lines), `indicator_manager.rs` (835) and `chart_layers_wiring.rs`
+# (270) are each under the 1,500-line threshold, so none takes an entry.
+# None of the three carries a test module, so those are production counts
+# entire.
+#
+# The budget arithmetic is done by hand, as the two branches' briefs agreed it
+# would be: `pane.rs` and `app.rs` were cut in parallel, both lowered this
+# line, and the second to land subtracts its own fall from the first's number
+# rather than from the one it measured against. 57,394 - 2,160 = 55,234.
+!budget 55234
 
-crates/app/src/app.rs 7499
+crates/app/src/app.rs 5339
 
 # The five entries below `pane.rs` were invisible to the first version of this
 # guard, which stopped counting at the first `#[cfg(test)]` of any kind:
