+++ docs/evidence/app-rs-workspace-and-indicators/report-diff.txt
@@ -1,0 +1,26 @@
+1c1
+< crate.lines.app	108188
+---
+> crate.lines.app	108269
+20c20
+< crate.lines.total	173375
+---
+> crate.lines.total	173456
+22c22
+< file.largest.crates/app/src/app.rs	7499
+---
+> file.largest.crates/app/src/app.rs	5339
+41,43c41,43
+< ratchet.size.budget	59547
+< ratchet.size.recorded	59547
+< ratchet.size.measured	173375
+---
+> ratchet.size.budget	57387
+> ratchet.size.recorded	57387
+> ratchet.size.measured	173456
+52,53c52,53
+< site.cfg_test	604
+< app.lines.without_egui	104495
+---
+> site.cfg_test	605
+> app.lines.without_egui	104574
