+++ docs/evidence/app-rs-workspace-and-indicators/tests-before-raw.txt
@@ -1,0 +1,30 @@
+test window_size_tests::a_requested_size_is_honoured_however_small_it_is ... ok
+test workspace_bundle::tests::a_recent_entry_is_labelled_by_the_name_the_trader_typed ... ok
+test workspace_bundle::tests::a_name_becomes_a_file_the_system_accepts_without_losing_the_words ... ok
+test window_size_tests::a_saved_sliver_reopens_on_something_the_trader_can_grab ... ok
+test widgets::tests::a_vector_icon_paints_the_letter_it_was_given ... ok
+test workspace_bundle::tests::a_registered_store_that_refuses_its_section_stops_the_whole_import ... ok
+test workspace_bundle::tests::revisiting_a_file_moves_it_up_instead_of_duplicating_it ... ok
+test workspace_bundle::tests::a_file_that_is_not_a_workspace_is_refused_when_it_is_opened ... ok
+test workspace_bundle::tests::the_recent_list_stays_a_menu ... ok
+test workspace_store::tests::a_blocked_store_never_asks_to_write ... ok
+test workspace_store::tests::a_change_inside_the_debounce_window_is_not_yet_asked_for ... ok
+test workspace_bundle::tests::an_entry_whose_file_is_gone_leaves_the_menu_but_not_the_list ... ok
+test ui_state::tests::the_history_reach_survives_a_save_and_a_reload ... ok
+test workspace_store::tests::a_change_that_has_settled_is_asked_for_once ... ok
+test workspace_store::tests::a_tab_switch_rebaselines_the_layers_rather_than_recording_a_switch ... ok
+test workspace_store::tests::the_exit_flush_ignores_the_debounce_but_not_the_absence_of_a_change ... ok
+test workspace_store::tests::marking_a_change_moves_the_flag_and_the_clock_together ... ok
+test workspace_store::tests::settling_a_replaced_book_overrides_what_landing_it_marked ... ok
+test workspace_bundle::tests::a_store_this_module_never_mentions_travels_on_registration_alone ... ok
+test workspace_bundle::tests::exporting_an_unreadable_store_fails_rather_than_writing_a_dead_bundle ... ok
+test workspace_bundle::tests::the_paper_sidecar_is_not_part_of_a_workspace ... ok
+test workspace_bundle::tests::a_store_that_was_never_written_is_simply_absent ... ok
+test workspace_bundle::tests::a_bundle_with_one_bad_section_writes_nothing_at_all ... ok
+test workspace_bundle::tests::a_saved_cockpit_comes_back_whole ... ok
+test workspace_bundle::tests::a_section_from_a_store_this_build_lacks_is_skipped_not_fatal ... ok
+test workspace_bundle::tests::the_facts_about_this_machine_never_travel_and_are_never_replaced ... ok
+test paper_trading::tests::a_tag_offers_its_cancel_exactly_while_it_paints_one ... ok
+
+test result: ok. 1894 passed; 0 failed; 4 ignored; 0 measured; 0 filtered out; finished in 6.78s
+
